ROS2 package + geometry refinement (#18)
* Updating gitignore for a ros development in ros2_node folder

* Hello lh_tracker_msgs package, it contains definition for LHtracker msg

* Initial definition of lh_tracker package, including:
- lh_tracker_serial node,
- lh_traker_geometry node.

* Hello lh_tracker_serial node:
The goal of this node is to receive data directly from the FPGA, parse 
them and if they seems valid, send them to lh_tracker_geometrical node 
for a geometrical analysis of this situation.

* Probably one of the most important part and complex part of this node:
Here's the goal of this class:
Given a set of two iterations retrieved by the FPGA hardware, retrieve 
the azimuth and elevation of the line which intersects the center of the 
lighthouse and the given sensor.
In order to work properly and accurately, some factory calibration data 
need to be filled up in this class:
- tilt_1: The physical inclination of the first plane relative to the Z 
axis of the LH,
- tilt_2: The physical inclination of the second plane relative to the Z 
axis of the LH,
- phase : The physical angles between the two lenses of the LH,
- start_angle: The physical angle which represent the iteration 0 of the 
LFSR.

* Here's some constants over the system and hardware

* Hello lh_tracker_geometry node:
The goal of this node if to be the interface between lh_tracker_serial 
node and the rest of the global software of the robot.
It receives iterations from lh_tracker_serial node, determines the pose 
of the lh_tracker in "map" frame knowing initial pose of a tracker.
From this transformation, it can track whatever sensor position knowing 
they are in the same plane relative to "map" frame, by finding the 
intersection point between this plane and the line which can be 
recovered from iterations from FPGA (with iteration, the mathematical 
model of the LH that we determines returns azimuth and elevation of the 
line starting from the center of the LH to the sensor itself).

* Nominal localisation of a set of sensor:
Given a batch of position linked to the id of sensors, by knowing 
physical position of each sensor relative to its frame, we can easily 
determines the pose of the tracker and then send through /lh_est_pose

* Factory physical calibration refining
This part of the software aims at refining factory calibration data 
implemeted in the lh itself. In fact, lh tracker calibration data are 
acceptable for relative positionning but not for absolute. That's why it 
needed a bit more work in order to refine physical calibration.
It now reaches 5mm absolute precision, it was nearly 30mm before this 
refining.

* Launch file, launch both lh_tracker_serial and lh_tracker_geometry nodes

* Blacked

* removing coms